% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{agnitio_settings}
\alias{agnitio_settings}
\title{Background function which is used to check if a token is available and valid, and obtain one if not.}
\usage{
agnitio_settings()
}
\description{
Background function which is used to check if a token is available and valid, and obtain one if not.
}
