% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{agnitio_token}
\alias{agnitio_token}
\title{Background function which is used to take Emsi Agnitio login details set by \code{\link{EpistemeLogin}} and obtains a token.}
\usage{
agnitio_token()
}
\description{
Background function which is used to take Emsi Agnitio login details set by \code{\link{EpistemeLogin}} and obtains a token.
}
