% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{datapull}
\alias{datapull}
\title{Pull data from the Agnitio API}
\usage{
datapull(
  country,
  content,
  release,
  constraints,
  metrics,
  limiter = 20000,
  brake = 5
)
}
\arguments{
\item{country}{The two-character country code.}

\item{content}{The Emsi Agnitio dataset description (e.g. "Occupation").}

\item{release}{The release or version identifier for the dataset (e.g. "2016.1").}

\item{constraints}{A list of dimensional constraints relevant to the dataset, each of which has been prepared through
\code{\link{dimmaker}} or \code{\link{CoW}}.}

\item{metrics}{A set of metrics through which to quantify the data, prepared through \code{\link{metricmaker}}.}

\item{limiter}{(optional) numerical limit on the dimensional size of your query, by default limited to 20,000
as the product of all the groups you're seeking to download - setting a higher number can make this work,
but think about rationing or saving the result.}

\item{brake}{(optional) changes the buffering time between queries to the API - default is set at 5 seconds,
and if you're planning to hammer the API with lots of queries, it's good to have this at least, maybe more.}
}
\value{
A data frame of dimensions and metrics, with dimensions classified as factors and metrics as doubles.
}
\description{
Pulls data from the Emsi Agnitio API according to specified parameters and returns a prepared data frame for analysis.
}
\examples{
met1 <- data.frame(names=c("Jobs.2016","Jobs.2022"), as=c("Jobs.2016","Jobs.2022"))
met1 <- metricmaker(met1)
area1 <- data.frame(name=c("Great Britain", "Wales"), code=c("GB", "WAL"))
areadim <- dimmaker("Area", area1)
occdim <- dimmaker("Occupation", data.frame(code="1"))
datapull("UK","Occupation","2016.1",list(CoW("A"),areadim,occsdim),met1)
}
