% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{dataset_dimension}
\alias{dataset_dimension}
\title{Query the details of an Agnitio API dataset}
\usage{
dataset_dimension(country, content, release, dimension)
}
\arguments{
\item{country}{The two-digit country identifier for the dataset.}

\item{content}{The keyword identifier for the content of the dataset (e.g. \code{"Occupation"}, \code{"Industry"}).}

\item{release}{The release or version identifier for the dataset (e.g. \code{"2016.1"}).}

\item{dimension}{The dimension you wish to explore.}
}
\value{
A data frame with the hierarchy and categories.
}
\description{
Identify the available metrics and required dimensions for a dataset held on Emsi Episteme.
}
\examples{
dataset_detail("UK","Occupation","2016.1")
}
